+++ django_scylla/cql/query.py
@@ -1,7 +1,6 @@
 import logging
 
 from django.db.models import sql
-from django.db.models.sql.datastructures import BaseTable
 
 from django_scylla.cql.where import WhereNode
 
@@ -11,25 +10,12 @@ logger = logging.getLogger(__name__)
 class Query(sql.query.Query):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
+        self.select_related = False
         self.where = WhereNode()
 
     def setup_joins(self, *args, **kwargs):
-        join_info = super().setup_joins(*args, **kwargs)
-
-        def final_transformer(field, alias):
-            if not self.alias_cols:
-                alias = None
-            return field.get_col(alias)
-
-        join_info = sql.query.JoinInfo(
-            join_info.final_field,
-            join_info.targets,
-            join_info.opts,
-            [join_info.joins[0]],
-            [],
-            final_transformer,
-        )
-        return join_info
+        self.select_related = False
+        return super().setup_joins(*args, **kwargs)
 
     def clear_where(self):
         self.where = WhereNode()
@@ -39,17 +25,17 @@ class Query(sql.query.Query):
         return super().trim_start(names_with_path)
 
     def join(self, join, *args, **kwargs):
-        if isinstance(join, BaseTable):
-            return super().join(join, *args, **kwargs)
-        return [self.base_table]
+        join.join_type = None
+        self.select_related = False
+        return super().join(join, *args, **kwargs)
 
     def add_ordering(self, *ordering):
         if len(ordering) > 1:
             ordering = [ordering[0]]
         return super().add_ordering(*ordering)
 
-    def add_select_related(self, fields):
-        ...
+    def add_select_related(self, _):
+        self.select_related = False
 
     def exists(self, using, limit=True):
         q = self.clone()